Fix tests for when the session delegate queue is not the main queue
The only problem is that tests assume that notifications match the start/stop
state of fetchers as tracked by the service. When the delegate queue is a
background queue and notifications are delayed til the main thread (which may
fire much later) no longer is it possible for test code to know accurately
which fetchers are in flight and which have already finished.

These changes just remove assumptions that the notifications seen by the test
reflect the current state of the fetcher tracking arrays inside the service
object.