Make synchronization uniform in the fetcher and upload fetcher
This gets it nearly ready to change the delegate queue from main queue to a
serial background queue.

The trickiest aspects of thread safety in the fetcher are that beginFetch may
be invoked again in a retry (either an auth retry or auto retry), and that
calls into the fetcher (mainly stopFetching) may occur at any time.

Changes in this CL:
- Exposed the NSURLSession delegate queue as sessionDelegateQueue. It still
  defaults to the main queue.
- Added a GTMSessionMonitorSynchronized set of macros and utility object so the
  @synchronized state can be verified at method entry points. The checks
  currently enforce non-recursive syncs, though the macros do support recursive
  syncs (and track the names of all functions currently syncing on the current
  thread.)
- Changed notifications to be posted consistently on the main thread. I'd
  rather they were always posted async but that breaks some current tests, and
  may affect client apps.
- Added a getter for initialBeginFetchDate so tests can easily discard delayed
  notifications from older fetchers.
- Fixed callbackQueue declaration to be a resettable property, defaulting to
  main queue.
- Added "Unsynchronized" accessors for internal accessors to the response and
  status codes. (I'm debating making those simple accessors allow recursive
  synchronization to avoid the extra methods.)

Also, a bug fix: stopFetchReleasingCallbacks previously had a warning that when
cancelling a task, the session shouldn't be invalidated until
URLSession:task:didCompleteWithError: was invoked due to an OS crash, but it
was neglecting to invalidate the session in that delegate method, causing tests
to leak sessions. This CL adds a _sessionNeedingInvalidation ivar to track and
release those sessions.