Start addressing race condition/deadlock issue.
Address a race condition that could lead to crash, and fixes
synchronization/locking order to prevent a closely related deadlock.

The underlying race condition is in scheduling invalidation of shared
NSURLSessions after a period of inactivity, and the structure of the code
doing so.

NSURLSession creation is performed by the fetcher, and the resulting session is
passed back to the service (and through it to the delegate dispatcher). To
prevent multiple fetchers from simultaneously attempting to create a
NSURLSession, a semaphore is used. The order of
@synchronized/dispatch_semaphore_wait was not strictly observed, and the
discard timer firing at just the right time could cause a deadlock. This change
moves ownership of that semaphore from the delegate dispatcher to the session
fetcher service, and makes the order of @synchronized/dispatch_semaphore_wait
locking consistent, which will prevent the deadlock.

The other race condition issue relates to invalidation of shared NSURLSession
in the service/delegate dispatcher, versus how the fetcher creates
NSURLSessionTasks. When beginning a request (for a fetcher sharing its
NSURLSession), the fetcher asks the service for the shared session; if one
exists, it is returned. At that point, there is no locking preventing
invalidation of the NSURLSession, but the fetcher will attempt to create an
NSURLSessionTask with that NSURLSession; if the session is invalidated before
task creation, trying to create the task will fail or crash.

This change tweaks the timer-based reset of the session, taking advantage of
how the timer callback does not nil the timer pointer, whereas the
-startSessionUsage method called when starting a new fetch does. Before
resetting the session when the timer fires, the service locks the session
creation semaphore (to prevent fetchers from being provided the NSURLSession),
and checks if the timer pointer is still non-nil; if so, no fetch has begun
since the discard timer was started, and while the semaphore is locked the
NSURLSession can be safely invalidated (unless some other deadlock or delay
has prevented the fetcher from starting its task, which would probably be a
different bug).

This CL fixes the deadlock. And it alleviates the race condition between
NSURLSession invalidation vs NSURLSessionTask creation, although doesn't
entirely eliminate it; triggering it should now require an extremely long and
unusual delay between the fetcher receiving the session from
-sessionForFetcherCreation and creating its task, with other fetchers having
finished in the meantime, which for most cases should now be much harder, but
at least no worse. Fully eliminating the race may need to wait on refactoring
the service/fetcher interaction, moving session creation/maintenance and likely
task creation into the service.