Fix GTMSessionUploadFetcher hanging on retry.
Changes the retry flow to have retryFetch call the new method
beginFetchForRetry. The upload fetcher overrides this to do a little
bookkeeping, then call the superclass implementation. The superclass
implementation calls -beginFetchMayDelay:mayAuthorize:, avoiding the
need to re-pass the completion handler.

The problem was that if the upload fetcher's initial request failed and
attempted to retry, the retry mechanism was losing the client's original
completion handler. This was due to the retry starting in retryFetch,
which called beginFetchWithCompletionHandler: passing the
GTMSessionFetcher's completion handler back into itself. When using an
upload fetcher, beginFetchWithCompletionHandler: is overridden to store
the provided handler and call the superclass with a new completion
handler used to begin triggering the chunk uploads. When the retry
occured, this caused the completion handler originally passed to the
upload fetcher to be discarded, so the client never got called when the
upload completes or errors out.