Properly handle query for offset when file size is unknown. Does not consider the upload complete when the returned query offset is greater than fileSize if fileSize is unknown (kGTMSessionUploadFetcherUnknownFileSize).
Fixes the following issue:
If the file size is unknown (kGTMSessionUploadFetcherUnknownFileSize), and a "query" command is sent to Scotty, the offset returned by the query will be ignored as the upload is considered complete (offset > fileSize). This causes the next upload command to fail with mismatched_content_start because we try to upload at offset 0, when we should be at a greater offset.