Fix Test build due to MD5 deprecations.
In the iOS 13/Mac 10.15 SDK the CC_MD5() is officially deprecated, but was being used within tests (only) for a means of testing authentication mechanisms. Pending a more thorough review of whether this test should be rewritten, updating the test code to enable targeting the tests at iOS 13/Mac 10.15 targets.