Release fetcher's active background task while awaiting retry.
When a fetch fails and attempts to retry, the background task was not being ended until the retry timer fires. In cases where the retry time may be large, the task may not complete for a long time, potentially holding the task open with the system beyond the alloted execution time leading to the app being terminated by the system.

Before starting the retry timer, the background task should be released.