Check a service exists before marking created `NSURLSession` as shared. (#198)
The fetcher only "shares" its `NSURLSession` when being created via a `GTMSessionFetcherService`. If no service is set, but an authorizer is provided and needs to request authorization before it can continue the fetch, when `canSharedSession == YES` the fetcher's `NSURLSession` is released but not invalidated, which can lead to a leak due to the session retaining its delegate (the fetcher) and itself.