Ensure callbacks are dispatched serially. (#230)
Ensure callbacks are dispatched serially.

Guarantees callbacks are dispatched serially by creating an intermediate serial queue targeting the user-specified callback queue (which may or may not be serial), and dispatching all callbacks to this intermediate queue. This ensures, for example, the accumulate data block will never be run after the completion handler.