New APIs -[GTMSessionFetcherService addDecorator:] and -removeDecorator: (#248)
New APIs `-[GTMSessionFetcherService addDecorator:]` and `-removeDecorator:`.

Allows creators and clients of a `GTMSessionFetcherService` to create and add decorators to intercept each created fetcher at fetch begin or retry time, and which can be notified when the fetch completes/finishes. These decorators may alter the `NSURLRequest` for the fetcher, adding HTTP headers or (e.g.) URL query parameters on retry.